requirements/base.txt
argon2-cffi==21.3.0
celery==5.2.3
crispy-bootstrap5==0.6
Django==3.2.11
django-celery-beat==2.2.1
django-crispy-forms==1.14.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-rosetta==0.9.8
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
Markdown==3.3.6
psycopg2-binary==2.9.3
Pillow==9.0.0
redis==4.1.2
